insurance_rag_assistant_gemini/requirements.txt

fastapi
uvicorn
pinecone
google-genai
